Fix factorial
For some reason the definition of the factorial of 0 being 1
does not work when calculating it manually. I'm not too into
the math but it seems this was somehow decided so I included
it as a special case. I'm vaguely aware that there is a way to
calculate this with recursion, but at this point in time
I would not like to challenge this.